bug: it fixes the bug appeared when an admin creates a new user in the platform
