+++ app/controllers/admin/users_controller.rb
@@ -1,14 +1,14 @@
 module Admin
   class UsersController < AuthenticatedController
     before_action :authorize_user
-    before_action :set_user, only: %i[edit update]
+    before_action :set_user, only: %i[edit update destroy]
 
     def index
       @users = User.joins(:profile).order(:first_name)
     end
 
     def new
-      @user = CreateUser.new
+      @user = Admin::CreateUser.new
     end
 
     def edit
@@ -17,13 +17,13 @@ module Admin
     end
 
     def create
-      @user = CreateUser.new user_params
+      @user = Admin::CreateUser.new create_user_params
 
       respond_to do |format|
         if @user.save
           format.html { redirect_to admin_users_path, notice: 'El usuario ha sido creado exitosamente' }
         else
-          format.turbo_stream { render turbo_stream: turbo_stream.replace('admin_user_form', partial: 'admin/users/form', locals: { user: @user }) }
+          format.turbo_stream { render turbo_stream: turbo_stream.replace('new_admin_user_form', partial: 'admin/users/forms/new_user', locals: { user: @user }) }
         end
       end
     end
@@ -53,7 +53,12 @@ module Admin
     end
 
     def destroy
-      debugger
+      session[:user_id] = nil if @user.eql? current_user
+      @user.destroy
+
+      respond_to do |format|
+        format.html { redirect_to admin_users_path, notice: 'El usuario ha sido eliminado exitosamente' }
+      end
     end
 
     private
@@ -67,8 +72,8 @@ module Admin
       authorize current_user
     end
 
-    def user_params
-      params.require(:admin_user).permit(:email, :password, :first_name, :last_name, :phone_number, :role)
+    def create_user_params
+      params.require(:admin_create_user).permit(:email, :password, :first_name, :last_name, :phone_number, :role)
     end
 
     def admin_update_account_params
