+++ app/views/admin/users/_user.html.erb
@@ -15,6 +15,8 @@
     <%= link_to edit_admin_user_path(user.token_id), class: 'flex items-center', data: { turbo_frame: 'admin_edit_user' } do %>
       <i class="fa-solid fa-pen-to-square text-primary"></i>
     <% end %>
-    <i class="fa-solid fa-trash text-danger"></i>
+    <%= button_to admin_user_path(user.token_id), method: :delete, class: 'flex items-center cursor-pointer', form: { data: { turbo_confirm: 'Estas seguro de eliminar a este usuario?' } } do %>
+      <i class="fa-solid fa-trash text-danger"></i>
+    <% end %>
   </div>
 </div>
