+++ app/views/admin/users/forms/_new_user.html.erb
@@ -1,0 +1,60 @@
+<%= form_with(model: user, url: admin_users_path, method: :post, class: 'flex flex-col w-full gap-5 items-center', id: 'new_admin_user_form') do |f| %>
+  <div class="flex items-start gap-2 w-full">
+    <div class="flex flex-col w-1/2 items-start">
+      <%= f.label :email, 'Correo electrónico:', class: 'font-bold' %>
+      <%= f.email_field :email, class: "input-tailwindcss font-semibold #{ error_input_border(user.errors.include?(:email)) }" %>
+      <% if user.errors.include?(:email) %>
+        <%= render 'shared/input_error_msg', { object: user, attribute: :email } %>
+      <% end %>
+    </div>
+
+    <div class="flex flex-col w-1/2 items-start">
+      <%= f.label :password, 'Contraseña:', class: 'font-bold' %>
+      <%= f.password_field :password, class: "input-tailwindcss font-semibold #{ error_input_border(user.errors.include?(:password)) }" %>
+      <span class="text-sm text-sidebarTextColor font-medium italic">La contraseña debe tener al menos 8 caracteres</span>
+      <% if user.errors.include?(:password) %>
+        <%= render 'shared/input_error_msg', { object: user, attribute: :password } %>
+      <% end %>
+    </div>
+  </div>
+
+  <div class="flex items-start gap-2 w-full">
+    <div class="flex flex-col gap-1 w-1/2 items-start">
+      <%= f.label :first_name, 'Nombre:', class: 'font-bold mb-1' %>
+      <%= f.text_field :first_name, class: "input-tailwindcss font-semibold #{ error_input_border(user.errors.include?(:first_name)) }" %>
+      <% if user.errors.include?(:first_name) %>
+        <%= render 'shared/input_error_msg', { object: user, attribute: :first_name } %>
+      <% end %>
+    </div>
+
+    <div class="flex flex-col gap-1 w-1/2 items-start">
+      <%= f.label :last_name, 'Apellido(s):', class: 'font-bold mb-1' %>
+      <%= f.text_field :last_name, class: "input-tailwindcss font-semibold #{ error_input_border(user.errors.include?(:last_name)) }" %>
+      <% if user.errors.include?(:last_name) %>
+        <%= render 'shared/input_error_msg', { object: user, attribute: :last_name } %>
+      <% end %>
+    </div>
+  </div>
+
+  <div class="flex items-start gap-2 w-full">
+    <div class="flex flex-col gap-1 w-1/2 items-start">
+      <%= f.label :phone_number, 'Número de teléfono:', class: 'font-bold mb-1' %>
+      <%= f.text_field :phone_number, class: "input-tailwindcss font-semibold #{ error_input_border(user.errors.include?(:phone_number)) }" %>
+      <% if user.errors.include?(:phone_number) %>
+        <%= render 'shared/input_error_msg', { object: user, attribute: :phone_number } %>
+      <% end %>
+    </div>
+
+    <div class="flex flex-col w-1/2 gap-1 items-start">
+      <%= f.label :role, 'Rol:', class: 'font-bold mb-1' %>
+      <%= f.select(:role, User.roles.keys.map { |n| [humanize_role(n), n] }, { include_blank: true },{ class: "input-tailwindcss font-semibold #{ error_input_border(user.errors.include?(:role)) }" }) %>
+      <% if user.errors.include?(:role) %>
+        <%= render 'shared/input_error_msg', { object: user, attribute: :role } %>
+      <% end %>
+    </div>
+  </div>
+
+  <div class="flex flex-col w-full items-center mt-10">
+    <%= f.submit "Crear cuenta", class: 'text-white py-3 rounded-lg shadow bg-black duration-300 hover:bg-gray-700 cursor-pointer font-semibold w-full text-center', data: { turbo_frame: '_top' } %>
+  </div>
+<% end %>
