+++ app/views/admin/users/new.html.erb
@@ -1,5 +1,5 @@
 <%= turbo_frame_tag 'admin_new_user', class:"flex flex-col w-full" do %>
   <%= render 'shared/modal', { title: 'Nuevo usuario' } do %>
-    <%= render 'form', user: @user %>
+    <%= render 'admin/users/forms/new_user', user: @user %>
   <% end %>
 <% end %>
