+++ app/views/shared/_input_error_msg.html.erb
@@ -1,3 +1,3 @@
-<span class="text-sm font-semibold text-danger">
+<span class="text-sm font-semibold text-danger break-all">
   <%= object.errors.full_messages_for(attribute).first %>
 </span>
